+++ gearlink/utils/__init__.py
@@ -3,6 +3,12 @@
 只有从此处显式导出的名称才是 utils 的公共 API。
 """
 
+from gearlink.utils.logging import disable_logging, enable_logging
 from gearlink.utils.token_count import count_message_tokens, estimate_tokens
 
-__all__ = ["estimate_tokens", "count_message_tokens"]
+__all__ = [
+    "estimate_tokens",
+    "count_message_tokens",
+    "enable_logging",
+    "disable_logging",
+]